+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.1] - Change API
+
+Change API:
+
+old: send(String controller, Map data, {Function listen})
+new: send({String controller, Map data, Function listen})
+
 ## [0.1.0] - Update readme
 
 ## [0.0.2] - Make Features